chore: split common exceptions into a package
Move OrchestratorExceedMaxIterationsException, InvalidToolCallParameterException,
ToolTimeoutError, TOOL_TIMEOUT_PHRASE, and ToolInitializationException into
src/quickapp/common/exceptions/, one class per file. Normalize each class on
super().__init__(message) and drop the redundant self.message+__str__ pairs
(keeping the .message attribute where external code reads it).
