+++ src/quickapp/common/exceptions/invalid_tool_call_parameter.py
@@ -1,0 +1,6 @@
+class InvalidToolCallParameterException(ValueError):
+
+    def __init__(self, parameter_name: str, message: str):
+        super().__init__(message)
+        self.parameter_name = parameter_name
+        self.message = message
