+++ src/quickapp/common/exceptions/orchestrator_exceed_max_iterations.py
@@ -1,0 +1,4 @@
+class OrchestratorExceedMaxIterationsException(RuntimeError):
+
+    def __init__(self):
+        super().__init__("Agent stopped due to max iterations.")
