+++ src/quickapp/common/exceptions/tool_initialization.py
@@ -1,5 +1,5 @@
 class ToolInitializationException(Exception):
-    """Exception raised when a tool or toolset fails to initialize properly."""
+    """Raised when a tool or toolset fails to initialize properly."""
 
     def __init__(
         self, message: str, tool_name: str = "", toolset_name: str = "", details: str = ""
@@ -7,7 +7,7 @@ class ToolInitializationException(Exception):
         super().__init__(
             f"Initialization Error: {message}" + (f" for Tool: {tool_name}" if tool_name else "")
         )
-        self.toolset_name = toolset_name
         self.message = message
         self.tool_name = tool_name
+        self.toolset_name = toolset_name
         self.details = details
